// app/shared/interfaces.ts
export interface IForm {
    opened_by: string;
    requested_by: string;
    cmdb_ci: string;
    risk: number;
    assignment_group: string;
    assigned_to: string;
    description: string;
    short_description: string;
    justification: string;
    implementation_plan: string;
    risk_impact_analysis: string;
    test_plan: string;
    u_communication_plan: string;
    start_date: string;
    state:string;
}